Fix deleted row handling
Add a number of missing checks against performing operations on
deleted lists, or deleted list entries, or entries of deleted lists.

The old logic tried to avoid the extra SELECT needed for those
checks by cramming all the check logic into the INSERT/UPDATE/DELETE
conditions and only doing explicit checks when the write failed.
This was unwieldy and the performance gain was unlikely to be
significant, so it is now replaced with explicit checks.

Bug: T177853
Change-Id: If8db3dee951be7038bb104eaf3a7c9f58b5f1723