Limit paging size
Per the DBA recommendation in T164990#3264314 / T168984#3659998,
limit paging sizes to 10 lists / 100 entries at a time.

For most queries this limits the number of rows scanned as well.
getListEntriesByDateUpdated is the only exception: it also scans
entries of deleted lists, which could result in an arbitrary
number of rows scanned. Fixing that is left to a separate patch.

Change-Id: I7c144fabcf5d7510a6b81ca85e623705daf01ccd