Unbreak query APIs
Apparently a null value for MAX2 works with the API help
but errors out when actually using the API. Setting it to the same
as MAX doesn't work well in help (will say something like
"limit is 10 (10 for bots)" which looks silly) but actually works.
That should probably be fixed upstream but for now let's go with
the latter.

Change-Id: Iee3f28cc9ca014c04461294ec5e9619365dac8c5