Provide timestamp to continue sync from
The readinglists and readinglistentries APIs take a timestamp
to list events from. Clients need a way to determine what timestamp
to use on their next sync call; that decision can be nontrivial
due to issues with multiple changes in the same second, transactioned
changes that are committed significantly later than inserted etc.
To avoid every client having to reimplement that logic, the API
will provide these timestamps.

Only the readinglists API provides such a timestamp for now, on the
assumption that clients will always call both APIs, and only want to
store a single "all changes included until" timestamp to simplify
internal state, so they can just call the readinglists API first and
use that timestamp for the list entry state as well.

Bug: T182706
Change-Id: I8247148ba5e58014754ae09bbe20905ae4a2b2e6