Decrement rl_size on entry deletion only if it is currently positive
Reading list sizes are tracked by an rl_size column which is meant to
stay in sync with the number of entries associated with the list in
reading_list_entries. Occasionally (i.e., for about 1 in 10000 lists,
absent intervention) it goes out of sync, and in such cases it
sometimes happens that decrementing the value of rl_size would cause
it to become negative; this causes an error because the column type is
unsigned int.

A new test is added for this behavior; I've also updated an existing
test because adding an already-deleted entry with addListEntries
caused the parent list's rl_size value to be 1 higher than expected.
The test is updated so that the rl_size value is consistent with
expectations, and rl_size is checked after entry deletion for this
list as well.

Bug: T259740
Change-Id: Ic66773fd539242734d93757a3ddaedf7c6bbff09