Don't make unnecessary API call(s) for anonymized reading list preview.
When anonymized preview is enabled, we're hiding everything including
the reading list title, description, and the actual list of articles,
but yet we're still making an API call that gets the details of all the
articles. This API call doesn't need to be made in this case.

Change-Id: I01679e8485081b9f54bf369a34e830a0f86384c4
(cherry picked from commit 20c47e632f2411a30c4bab132818c9e1b74d60ba)