Introduce Import button for launching deeplink into app.
This adds an "Import" button (only shown when sharing/browsing from
native devices) that launches the same URL but with a different
scheme ("wikipedia://") which will automatically invoke the app and
provide it with the same URL parameters from the current URL.

This will be a slightly less confusing workflow than asking the user
to go back to the original messaging app that received the shared link.
Instead, the user will install the app, then tap Back once, then tap
the Import button.

Bug: T313269
Change-Id: Id0d3ae0509f7c37a6f39770716f50f65cae41445