Fix importing / Enable viewing of watchlist to everyone
* Expose watchlist
This makes the reading list watchlist accessible to everyone.
The page shows the watchlist title when watchlist is visited

This is needed to remove Special:EditWatchlist in MobileFrontend
and gives it usage in production on the long run which will help
with code and feature quality.

* Restore display of imported list tab
Fixes bug introduced by the tab addition to the reading
list feature.

Bug: T266065
Change-Id: I55f55b2545cfaac02b33776bc1629d19bf5a5c02