Fix testing errors due to double-initialization of handler instance
Pagination tests were reusing the handler instance for the second
page of data. This was previously harmless, if inadvisable. But the
utility function to invoke the handler also initializes it, and a
recent core change makes double-initialization an error. Instead,
recreate the handler instance for the second call. This is more
respresentative of the real situation anyway, where each api call
would cause a separate handler instance to be created.

Change-Id: I85c2bd38a0a8ad61a6331e26f936d573cef7c939