Implements delete item from list functionality
* Adds translations
* Adds notification message when action is taken,
defensive programming checks if msg is missing
* Jest: add support for testing Vue components and increase
coverage to reflect new tests
* tests: Added list entries and delete button
* eslint: bumped ecmaVersion to support use of async functions,
which is supported according to the browser matrix.
* ReadingListPage now passes listId to ReadingList to support
notification messages

Additional change:
* Adds readinglist.page.icons module in preparation for future
task around adding/removing items from a list
* Fixes an issue with list entry order in the situation where
titles and redirects are normalized - the new order
of results is based on query.normalized instead of the title
* Fixes an issue where if lists contained two pages which
redirected to the same page, the image was lost
* This fixes an issue where the reading list entries are sometimes out of order
when there are multiple projects, as well as other edge cases.
* Test case: reading list entry IDs are not being preserved across projects

Bug: T389437
Change-Id: I0e199db11a409f7eb429fb00a53c0f5c145bcd71