Feature: add/remove page to saved list from article
The bookmark icon is shown on article pages. It is visibility hidden by default
until the API has checked to see if the article exists in the default list.
This also calls setup collections if needed.
A FIXME is added to clarify that this needs to be addressed at a later date.

Functional changes:
* bookmark/unbookmark action messages now support additional parameters for list
ID and name.
* `addPageFromReadingList` and `removePageFromReadingList` will now pass the listId to
setBookmarkStatus in order to provide the 3rd parameter for
messages`readinglists-browser-add-entry` and `readinglists-browser-remove-entry-success`.

API changes:
* API is pulled out into dedicated readinglist.api module so that it can be used
in  readinglist.page.scripts.js without needing to load Vue on page load.
* Implements create/deleteEntry functions (with tests)
* Updates all mw.api calls to use `formatversion: 2`
* Cache the `getDefaultReadingList()` return value to a local variable, so that
it is only called once

Additional changes:
* Removes unnecessary eslint overrides in jest folder

Bug: T386669
Change-Id: I6a03f825a70a36b2f56765ca3b8a60129bc29a22