Retain project and thumbnail for different projects
Fixed mismatched page thumbnails in API

This was a regression introduced by the merge request that
fixed mismatched IDs. Before it was only matching results by title,
but now it checks the project too.

Adds test case to protect against this in future.

Additional changes:
* lint: Fix a max-len warning
* Bump code coverage
* format: Just some test-related cosmetic changes: Nothing
important or fancy, this doesn't change any sort of functionality.

Bug: T389826
Change-Id: I5d1fc4dca633484b3b7f85451d563d88205db5a4