ReadingLists REST: add patch version to module definition file
We are establishing a convention where REST module versions should
include full semantic version information, like:
  major.minor.patch
ReadingLists previously did not specify a patch version, so add it.

This intentionally does not affect the "v0" portion of the moduleId,
just the "version" value. See T392149 for more details, or T395713
for info on just this portin of the overall plan.

Bug: T395713
Change-Id: Id793965ea99c18272c239e08140364fc0a980332