Reading Lists: add missing "entries" values to /changes/since response
The /changes/since endpoint should send a response continuing two
sets of values: "lists" and "entries". In the RESTBase+Action API
implementation, RESTBase queried these separately and combined them
into the response. This was missed in the conversion to MW REST API,
which returned only the "lists" values. Add the "entries" values
to correct this.

Also, disallow "name" as a sort value for the /changes/since endpoint.
To match the Action API implementation, and because sorting by name in
this case is nonsensical, this should never have been allowed. No known
existing callers will be affected.

Bug: T351149
Change-Id: I25b0b6dad6c10c7c35648aa86258625ee7654427