Script to update reading list projects urls
Wikis on the beta cluster have moved to wmcloud.org from
wmflabs.org.

The reading_list_project table uses the wiki canonical url
as an identifier for the project.  With the new domain,
adding a page to a reading list no longer works since the new domain
does not match anything in the reading_list_project table.

Bug: T400364
Change-Id: I85ea1d6d860572409f49d132de0b2cbb7460beef