tests: Fix ReadingLists teardown order
Do readingListsTeardown before MediaWiki (parent) tearDown

With I8a9718548f36c5a04441e23988aeecbefc4c70a6 in core,
it calls $session->persist() for the api tokens and needs
proper cleanup in the tests.

If we do the parent (core) tearDown first, then the token
is not valid for the extension cleanup.

If we swap the order, then the tearDown works properly.

Bug: T404229
Change-Id: I2c2f3b5df3cd0f6d73aa31e2fd85d02b89dd055e