Fix error when local user is not connected to CentralAuth
The error results in the page not rendering wherever
the bookmark icon button is present.

ReadingLists require users are connected to a central lookup,
since ReadingLists include pages from multiple wikis.

If the local user is not connected, then they are not eligible
for ReadingLists and they can continue to see the watchlist star.

Bug: T407365
Depends-On: I9a58d5840f4e9b73b4354bb0e4aa121e3fcc14b4
Change-Id: I59c3cbf2ae4c451499c1785aace78b2f07b96c44