i18n: Remove quotes from messages that mention bold page names
The page names in these messages are already highlighted
using <b>, so the double quotes are not so necessary.

In addition, quite a lot of articles have double quotes
in them, such as [["Heroes" (David Bowie song)]], and
putting them in double quotes again will look weird.
Many more examples:
https://en.wikipedia.org/wiki/Special:PrefixIndex?prefix=%22&namespace=0&hideredirects=1

Change-Id: I39a63f344986c18701adb75f4204ba55c252f641